requirements.txt
mobase-stubs
